Split mergeMemorySources script into focused modules (488 → 42 lines)
The legacy ChatGPT memory importer was one monolithic script with
five concerns interleaved. Each now lives in its own file under
scripts/merge-memory/:

- types.ts        shared types + path constants (REPO_ROOT now
                  resolves from this file's location, since the
                  modules live one level deeper than the entry point)
- text-utils.ts   small pure helpers: pathExists, normalizeWhitespace,
                  toIsoTime, shortHash, sanitizeFileName,
                  summarizePreview, extractTextFromContent, readJsonFile
- normalize.ts    single-conversation normalization + cross-source
                  mergeConversation (the de-dup pass that combines
                  the same conversation seen from two sources)
- sources.ts      disk discovery for the legacy export roots —
                  buildSourceManifests + the breadth-first
                  strategic-doc scan
- artifacts.ts    writeFoundationDocs + buildMergedConversations +
                  the markdown overview + writeArtifacts (the five
                  output writers)

mergeMemorySources.ts is now a 42-line main() that calls into the
modules in order. No behavior change — the artifacts produced are
byte-for-byte identical.
